SWC-6316
- Fix issue (likely import cycle)
